+++ paimon-core/src/test/java/org/apache/paimon/mergetree/compact/LookupChangelogMergeFunctionWrapperTest.java
@@ -26,7 +26,6 @@ import org.apache.paimon.lookup.LookupStrategy;
 import org.apache.paimon.mergetree.compact.aggregate.AggregateMergeFunction;
 import org.apache.paimon.mergetree.compact.aggregate.FieldAggregator;
 import org.apache.paimon.mergetree.compact.aggregate.FieldSumAgg;
-import org.apache.paimon.options.Options;
 import org.apache.paimon.types.DataField;
 import org.apache.paimon.types.DataTypes;
 import org.apache.paimon.types.IntType;
@@ -37,7 +36,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -370,30 +368,5 @@ public class LookupChangelogMergeFunctionWrapperTest {
         assertThat(changelogs).hasSize(0);
         kv = result.result();
         assertThat(kv).isNull();
-    }
-
-    @Test
-    public void testPartialUpdateIgnoreDelete() {
-        Options options = new Options();
-        LookupChangelogMergeFunctionWrapper function =
-                new LookupChangelogMergeFunctionWrapper(
-                        LookupMergeFunction.wrap(
-                                PartialUpdateMergeFunction.factory(
-                                        options,
-                                        DataTypes.ROW(DataTypes.INT()),
-                                        Collections.singletonList("f0")),
-                                RowType.of(DataTypes.INT()),
-                                RowType.of(DataTypes.INT())),
-                        key -> null,
-                        EQUALISER,
-                        false,
-                        LookupStrategy.CHANGELOG_ONLY,
-                        null);
-
-        function.reset();
-        function.add(new KeyValue().replace(row(1), 1, DELETE, row(1)).setLevel(2));
-        ChangelogResult result = function.getResult();
-        assertThat(result).isNotNull();
-        assertThat(result.result()).isNull();
     }
 }
