+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/TableActionBase.java
@@ -18,13 +18,11 @@
 
 package org.apache.paimon.flink.action;
 
-import org.apache.paimon.CoreOptions;
 import org.apache.paimon.catalog.Catalog;
 import org.apache.paimon.catalog.Identifier;
 import org.apache.paimon.flink.sink.FlinkSinkBuilder;
 import org.apache.paimon.table.FileStoreTable;
 import org.apache.paimon.table.Table;
-import org.apache.paimon.utils.Preconditions;
 
 import org.apache.flink.api.dag.Transformation;
 import org.apache.flink.streaming.api.datastream.DataStream;
@@ -35,7 +33,6 @@ import org.apache.flink.table.data.RowData;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -97,25 +94,6 @@ public abstract class TableActionBase extends ActionBase {
                     "Failed to invoke 'TableEnvironmentImpl#executeInternal(List, List)' method "
                             + "from given StreamTableEnvironment instance by Java reflection. This is unexpected.",
                     e);
-        }
-    }
-
-    /**
-     * The {@link CoreOptions.MergeEngine}s will process -U/-D records in different ways, but we
-     * want these records to be sunk directly. This method is a workaround. Actions that may produce
-     * -U/-D records can call this to disable merge engine settings and force compaction.
-     */
-    protected void changeIgnoreMergeEngine() {
-        if (CoreOptions.fromMap(table.options()).mergeEngine()
-                != CoreOptions.MergeEngine.DEDUPLICATE) {
-            Map<String, String> dynamicOptions = new HashMap<>();
-            dynamicOptions.put(
-                    CoreOptions.MERGE_ENGINE.key(), CoreOptions.MergeEngine.DEDUPLICATE.toString());
-            // force compaction
-            dynamicOptions.put(CoreOptions.FULL_COMPACTION_DELTA_COMMITS.key(), "1");
-            Preconditions.checkArgument(
-                    table instanceof FileStoreTable, "Only supports FileStoreTable.");
-            table = ((FileStoreTable) table).internalCopyWithoutCheck(dynamicOptions);
         }
     }
 }
